requirements.txt
google-generativeai # or the latest version
google-auth
google-api-core
requests
